backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
python-chess
celery[redis]
psycopg[binary]
pydantic
python-chess
celery
sqlmodel
psycopg[binary]
alembic
sse-starlette
sseclient-py
requests
pandas
numpy
scikit-learn